refactor: Spinner => MapSpinner 파일 이름 변경
