+++ src/pages/Map/components/MapComponent.tsx
@@ -3,9 +3,9 @@ import { useQuery } from '@tanstack/react-query';
 import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { Container as MapDiv, NaverMap, useNavermaps } from 'react-naver-maps';
 import GardensContainer from './GardensContainer';
+import MapSpinner from './MapSpinner';
 import GardenMarker from './Marker/GardenMarker';
 import MyMarker from './Marker/MyMarker';
-import MapSpinner from './Spinner';
 import useGeolocation from '@/hooks/useGeolocation';
 import gardensApi from '@/services/gardens/api';
 import { gardensQuery } from '@/services/gardens/query';
